feat: refine Confluence
