+++ src/apis/confluence/__init__.py
@@ -1,0 +1,11 @@
+"""
+API clients for Confluence.
+
+https://developer.atlassian.com/cloud/confluence/rest/intro/
+"""
+
+from .confluence import ConfluenceConnector
+
+__all__ = [
+    "ConfluenceConnector",
+]
